Current Refactor State
