+++ go.mod
@@ -3,6 +3,6 @@ module bey/go-vct
 go 1.21.1
 
 require (
-	github.com/joho/godotenv v1.5.1 // indirect
-	github.com/mattn/go-sqlite3 v1.14.22 // indirect
+	github.com/joho/godotenv v1.5.1
+	github.com/mattn/go-sqlite3 v1.14.22
 )
